REFACTOR: Moved constants to directory called `database` - These items emulate entities in a database
