+++ actions/skills/groupSkills.ts
@@ -1,5 +1,3 @@
-import { languages } from "@/constants/languages";
-import { technologies } from "@/constants/technologies";
 import Skill from "@/types/skills";
 
 const groupSkills = (
