+++ app/(site)/components/AboutSection/LanguageSection.tsx
@@ -3,9 +3,9 @@
 import Tag from "@/components/Tags/Tag";
 import LanguageModal from "@/components/Modal/LanguageModal";
 import HeadingThree from "@/components/Text/HeadingThree";
-import { languages } from "@/constants/languages";
 import Skill from "@/types/skills";
 import React, { useState } from "react";
+import { languages } from "@/database/skills/languages";
 
 /**
  * Displays a list of languages that I know.
