+++ app/(site)/components/AboutSection/SkillSection.tsx
@@ -4,10 +4,10 @@ import Tag from "@/components/Tags/Tag";
 import SkillsModal from "@/components/Modal/SkillsModal";
 import SkillTag from "@/components/Tags/SkillTag";
 import HeadingThree from "@/components/Text/HeadingThree";
-import { languages } from "@/constants/languages";
-import { technologies } from "@/constants/technologies";
 import Skill from "@/types/skills";
 import { useState } from "react";
+import { languages } from "@/database/skills/languages";
+import { technologies } from "@/database/skills/technologies";
 
 /**
  * Displays a list of skills that I have.
