+++ app/blogs/[slug]/page.tsx
@@ -3,7 +3,7 @@ import getMarkdownFromFileSystem from "@/actions/getMarkdownFromFileSystem";
 import Reader from "@/components/Reader/Reader";
 import SkillTableSection from "@/components/Skills/SkillTableSection";
 import HeadingTwo from "@/components/Text/HeadingTwo";
-import blogs from "@/constants/blogs";
+import blogs from "@/database/blogs";
 import Skill from "@/types/skills";
 import type { Metadata, ResolvingMetadata } from "next";
 import { notFound } from "next/navigation";
