+++ app/blogs/page.tsx
@@ -2,7 +2,7 @@ import HeadingOne from "@/components/Text/HeadingOne";
 import { BlogList } from "./components/BlogList";
 import type { Metadata } from "next";
 import PageDescription from "@/components/UI/PageDescription";
-import blogs from "@/constants/blogs";
+import blogs from "@/database/blogs";
 
 const description = `
   Explore my collection of blogs on various topics.
