+++ database/skills/generalSkills.ts
