+++ database/skills/hardSkills.ts
