+++ database/skills/skills.ts
@@ -1,0 +1,8 @@
+import Skill from "@/types/skills";
+import softSkills from "./softSkills";
+import hardSkills from "./hardSkills";
+import generalSkills from "./generalSkills";
+
+const allSkills: Skill[] = [...hardSkills, ...generalSkills, ...softSkills];
+
+export default allSkills;
