+++ database/skills/softSkills.ts
